requirements.txt
fastapi>=0.100.0,<1.0.0
uvicorn[standard]>=0.22.0,<1.0.0
pydantic>=2.0.0,<3.0.0
python-multipart>=0.0.6,<1.0.0
aiofiles>=23.1.0,<24.0.0
psutil>=5.9.0,<6.0.0
sqlalchemy>=2.0.0,<3.0.0
aiosqlite>=0.19.0,<1.0.0
python-dotenv>=1.0.0,<2.0.0
